backend/scripts/index_documents.py
import os
import logging
import glob
from dotenv import load_dotenv

# Document Loaders and splitters
from langchain_community.document_loaders import PyPDFLoader
from langchain_text_splitters import RecursiveCharacterTextSplitter

# azure components import 
from langchain_openai import AzureOpenAIEmbeddings
from langchain_community.vectorstores import AzureSearch

load_dotenv(override=True)

# setup logging
logging.basicConfig(
    level=logging.INFO,
    format = "%(asctime)s - %(levelname)s - %(message)s"
)

logger = logging.getLogger("indexer")

def index_docs():
    '''
    Reads the PDFs, chunks them and store them to Azure AI search
    '''

    # define the paths. we look for data folder
    current_dir = os.path.dirname(os.path.abspath(__file__))
    data_dir = os.path.join(current_dir,"../../backedn/data")

    # check the environment variables
    logger.info("="*60)
    logger.info("Environment Configuration Check")
    logger.info(f"AZURE_OPENAI_ENDPOINT : {os.getenv('AZURE_OPENAI_ENDPOINT')}")
    logger.info(f"AZURE_OPENAI_API_VERSION : {os.getenv('AZURE_OPENAI_API_VERSION')}")
    logger.info(f"Embedding Deployment : {os.getenv('AZURE_OPENAI_EMBEDDING_DEPLOYMENT','text-embedding-3-small')}")
    logger.info(f"AZURE_SEARCH_ENDPOINT:{os.getenv('AZURE_SEARCH_ENDPOINT')}")
    logger.info(f"AZURE_SEARCH_INDEX_NAME: {os.getenv('AZURE_SEARCH_INDEX_NAME')}")
    logger.info("="*60)

    required_vars = [
        "Azure_OPENAI_ENDPOINT",
        "AzURE_OPENAI_KEY",
        "AZURE_SEARCH_ENDPOINT",
        "AZURE_SEARCH_API_KEY",
        "AZURE_SEARCH_INDEX_NAME"
    ]

    missing_vars = [var for var in required_vars if not os.getenv(var)]
    
    if missing_vars:
        logger.error(f"Missing environment variables: {missing_vars}")
        logger.error("Please check your .env file and ensure that all enviornment variables are set")
        return
    
    try:
        logger.info("Initializing the Azure OpenAI Embeddings....")
        AzureOpenAIEmbeddings(
            azure_deployment=os.getenv('AZURE_OPENAI_EMBEDDING_DEPLOYMENT', 'text-embedding-3-small'),
            azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"),
            api_key=os.getenv("AZURE_OPENAI_API_KEY"),
            openai_api_version = os.getenv("AzURE_OPENAI_API_VERSION", "2024-02-01")
        )
    except Exception as e:
        logger.error(f"Failed to initialize the embeddings {e}")
        logger.error("Please verify your Azure Open AI deployment name and endpoint")
        return
                                  

    try:
        logger.info("Initializing the Azure AI Search Vector Store")
        embeddings = AzureOpenAIEmbeddings(
            azure_search_endpoint = os.getenv("AZURE_SEARCH_ENDPOINT"),
            azure_search_key = os.getenv("AZURE_SEARCH_API_KEY"),
            index_name = index_name,
            embedding_function = embeddings.embed_query            
        )
        logger.info(f"Vector store initialized for index {index_name}")
    except Exception as e:
        logger.error(f"Failed to initialize the Azure AI Search : {e}")
        logger.error("Please verify your Azure Search endpoint, API key and index name")
        return


    # find PDF files

    pdf_files = glob.glob(os.path.join(data_dir,"*.pdf"))

    if not pdf_files:
        logger.warning(f"No PDF Files were found in {data_dir}. Please add the files")
    logger.info(f"Found {len(pdf_files)} PDFs to process : {[os.path.basename(f) for f in pdf_files]}")

    all_splits = []
    for pdf_path in pdf_files:
        try:
            logger.info(f'Loading {os.path.basename(pdf_path)}....')
            loader = PyPDFLoader(pdf_path)
            raw_doc = loader.load()

            text_splitter = RecursiveCharacterTextSplitter(
                chunk_size = 1000,
                chunk_overlap = 200
            )

            splits = text_splitter.split_documents(raw_doc)
            for split in splits:
                split.metadata["source"] = os.path.basename(pdf_path)
            
            all_splits.extend(splits)
            logger.info(f"Split into {len(splits)} chunks")

        except Exception as e:
            logger.error(f"Failed to process {pdf_path}")


    # Upload to Azure

    if all_splits:
        logger.info(f"Uploading {len(all_splits)} chunks to Azure AI Search Index: {index_name}")
        try: 
            vector_store.add_documents(documents = all_splits)
            logger.info("=" * 60)
            logger.info("Indexing is complete !  Knowledge base is ready")
            logger.info(f"Total chuns indexed: {len(all_splits)}")
            logger.info("=" * 60)
        except Exception as e:
            logger.error(f" Failed to upload the documents to Azure Search {e}")
            logger.error(f"Please check the Azure Search configuration and try again")
    
    else:
        logger.warning("No Documents were processed")


if __name__ = "__main__":
    index_docs()


