backend/src/services/video_indexer.py
import os
import logging
import time
import requests
import yt_dlp
from azure.identity import DefaultAzureCredential

logger = logging.getLogger("video-indexer")

class VideoIndexerService:
    def __init__(self):
        self.account_id = os.getenv("AZURE_VI_ACCOUNT_ID")
        self.location = os.getenv("AZURE_VI_LOCATION")
        self.subscription_id = os.getenv("AZURE_SUBSCRIPTION_ID")
        self.resource_group = os.getenv("AZURE_RESOURCE_GROUP")
        self.vi_name = os.getenv("AZURE_VI_NAME", "brand-yt-vi")
        self.credential = DefaultAzureCredential()

    def get_access_token(self):
        '''
        Generate an ARM Access token

        '''
        try:
        
            token_object = self.credential.get_token("https://management.azure.com/.default")
            return token_object.token
        
        except Exception as e:
            logger.error(f"Failed to get Azure Token {e}")
            raise

    def get_account_token(self, arm_access_token):
        '''
        Exchanges the ARM token for video indexer account team

        '''
        url = (
            f"https://management.azure.com/subscriptions/{self.subscription_id}"
            f"/resourceGroups/{self.resource_group}"
            f"/providers/Microsoft.VideoIndexer/accounts/{self.vi_name}"
            f"/generateAccessToken?api-version=2024-01-01"
        )

        headers = {"Authorization": f"Bearer {arm_access_token}"}
        payload = {"PermissionType": "Contributor", "scope" : "Account"}
        response = requests.post(url, headers= headers, json=payload)

        if response.status_code!=200:
            raise Exception(f"Failed to get VI Account Token : {response.text}")
        return response.json().get("accessToken")

    # Function to download the youtube video

    def download_youtube_video(self,url, output_path = "temp_video.mp4"):
        '''
        downloads the youtube video to a local file
        '''
        logger.info(f"Downloading the youtube video : {url} ")

        ydl_opts = {
            "format" : 'best',
            'outtmpl': output_path,
            'quiet': False,
            'no_warnings': False,
            'extractor_args': {'youtube': {'player_client': ['android', 'web']}},
            'http_headers' : {
                'User-Agent' : 'Mozilla/5.0 (Windows NT 10.0; Win64; X64) AppleWebKit/537.36'
            }
        }

        try:
            with yt_dlp.YoutubeDL(ydl_opts) as ydl:
                ydl.download([url])
            logger.info("Download complete")
            return output_path
        except Exception as e:
            raise Exception(f"Youtube video download failed : {str(e)}")
        
    def upload_video(self, video_path, video_name):
        arm_token = self.get_access_token()
        vi_token = self.get_account_token(arm_token)

        api_url = f"https://api.videoindexer.ai/{self.location}/Accounts/{self.account_id}/Videos"
      
        params = {
            "accessToken" : vi_token,
            "name" : video_name,
            "privacy": "Private",
            "indexingPreset": "Default"
        }

        logger.info(f"Uploading the video {video_path} to Azure")

        # open the file in binary and stream it on azure

        with open(video_path,'rb') as video_file:
            files = {'file': video_file}
            response = requests.post(api_url,params=params,files=files)
        
        if response.status_code!=200:
            raise Exception(f"Azure Upload failed2 {response.text}")
        
    def wait_for_processing(self, video_id):
        while True:
            arm_token = self.get_access_token()
            vi_token = self.get_account_token(arm_token)
            url = f"https://api.videoindexer.ai/{self.location}/Accounts/{self.account_id}/Videos"
            params = {'accessToken': vi_token}
            response = requests.get(url,params=params)
            data = response.json()

            state = data.get("state")
            if state == "Processed":
                return data
            elif state =="Failed":
                raise Exception("Video indexing failed in Azure")
            elif state == "Quarantined":
                raise Exception("Video Quarantied (Copyright/Content Policy violation)")
            logger.info(f"Status {state} waiting for 30 sec")
            time.sleep(30)

    def extract_data(self, vi_json):
        transcript_lines = []
        for v in vi_json.get("videos",[]):
            for insight in v.get("insigts",{}).get("transcript",[]):
                transcript_lines.append(insight.get("text"))

        ocr_lines = []
        for v in vi_json.get("videos",[]):
            for inights in v.get("insights", {}).get("ocr",[]):
                ocr_lines.append(insight.get("text"))

        return {
            "transcript": " ".join(transcript_lines),
            "ocr_text": ocr_lines,
            "video_metadata": {
                "duration": vi_json.get("summarizedinsights",{}).get("duration"),
                "platform": "youtube"
            }
        }

